% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_contact.R
\name{contact_network}
\alias{contact_network}
\title{This function connects to the sormas db and generates the data specified
by issue https://github.com/hzi-braunschweig/SORMAS-Stats/issues/88
the data will be usfull for further statistics ralated to contacts}
\usage{
contact_network(sormas_db)
}
\description{
This function connects to the sormas db and generates the data specified
by issue https://github.com/hzi-braunschweig/SORMAS-Stats/issues/88
the data will be usfull for further statistics ralated to contacts
}
