% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sormas_db.R
\name{do_connect}
\alias{do_connect}
\title{Connects to a PostgreSQL DB}
\usage{
do_connect(db_name, user, pwd)
}
\arguments{
\item{db_name}{The name of the database}

\item{user}{The name of the database}

\item{pwd}{The name of the database}
}
\value{
A connection to the DB
}
\description{
Connects to a PostgreSQL DB
}
